Inject a letter in a set of sequences at random positions

Inject a letter in a set of sequences at random positions

inject_letter_random(my_seq, how_many_seqs = NULL, how_many_letters = NULL,
seed = NULL, letter = "N")

Arguments

my_seq
character vector with sequences to inject

how_many_seqs
How many sequences pick to inject Ns. An interval [min_s, max_s] with min_s minimum and max_s maximum sequences can be passed. In this case, a value is picked from the interval. If NULL, a random value within the interval [1, length(my_seq)] is picked.

how_many_letters
How many times inject the letter in the i sequences that are going to be injected. An interval [min_i max_i] can be passed. In this case, a value is randomly picked for each sequence i. This value represents the number of times that the letter will be injected in the sequence i. If NULL, a random value within the interval [1, width(my_seq[i])] is picked for each sequence i.

seed
Random seed. Default: NULL

letter
Letter to inject. Default: 'N'

Value

character vector

Examples

s <- random_seq(slength = 10, swidth = 20, seed = 10)
s <- inject_letter_random(s, how_many_seqs = 1:30, how_many= 2:10, seed = 10)