Remove left and right full and partial patterns

This program can remove adapters and partial adapters from 3' and 5'. The adapters can be anchored or not. Two methods are available: one based on the exact matching of the sequence and the adapter, and a second in an error rate. When indels are allowed, the error rate consists in the edit distance. IUPAC simbols are allowed. The methods use the isMatchingStartingAt function of the Biostrings package to find matches. IUPAC symbols are allowed in all the cases. The present function also removes partial adapters, without the need of additional steps (for example, creating a padded adapter with 'Ns', etc). A similar to what obtained with trimLRPatterns when the raw adaptor is passed, can be obtained with the option anchored = TRUE. When several matches are found, the function removes the subsequence starting from the first match in cutRseq, and ending in the last one in cutLseq.

cutRseq(subject, Rpattern, with.indels = FALSE, fixed = "subject",
method = c("exact", "er"), error_rate = 0.2, anchored = TRUE,
ranges = FALSE, checks = TRUE, min_match_flank = 0L, ...)

cutLseq(subject, Lpattern, with.indels = FALSE, fixed = "subject",
method = c("exact", "er"), error_rate = 0.2, anchored = TRUE,
ranges = FALSE, min_match_flank = 1L, checks = TRUE, ...)

Arguments

subject
DNAString or DNAStringSet object

Rpattern
3' pattern, DNAString object

with.indels
Allow indels? This is available only for er method.

fixed
Parameter passed to codeisMatchingStartingAt and isMatchingEndingAt Default 'subject', ambiguities in the pattern only are interpreted as wildcard. See the argument fixed in codeisMatchingStartingAt and isMatchingEndingAt

method
Method used for trimming. If 'exact' the metod is based in the exact matching of the subsequences of the subject and adapters If 'er' the metod is based on the eror-rate between the subsequences, allowing mismatches.

error_rate
Error rate (value in [0, 1] value for 'er' method. The error rate is the proportion of mismatches allowed between the adapter and the aligned portion of the subject. For a given adapter A, the number of allowed mismatches between each subsequence s of A and the subject is computed as error_rate * L_s, where L_s is the length of the subsequence s.

anchored
Can the adapter or partial adapter be within the sequence? (dafault: anchored = TRUE) or only in the terminal regions of the sequence (anchored; anchored = TRUE). Default TRUE (trim only flanking regions).

ranges
return ranges? Default FALSE

checks
perform internal checks? Default TRUE

min_match_flank
Do not trim in flanks of the subject, if a match has min_match_flank of less length. Default 1L (only trim with >=2 coincidences in a flank match)

...
additional parameters passed to isMatchingStartingAt and isMatchingEndingAt

Lpattern
5' pattern, DNAString object

Value

Edited DNAString or DNAStringSet object

Examples

library(Biostrings)

subject <- DNAStringSet(c('ATCATGCCATCATGAT',
'CATGATATTA', 'TCATG', 'AAAAAA', 'AGGTCATG'))

Lpattern <- Rpattern <- 'TCATG'

cutLseq(subject, Lpattern)

#>   A DNAStringSet instance of length 5
#>     width seq
#> [1]    16 ATCATGCCATCATGAT
#> [2]     6 ATATTA
#> [3]     0 
#> [4]     6 AAAAAA
#> [5]     8 AGGTCATG

cutLseq(subject, Lpattern, ranges = TRUE)

#> IRanges object with 5 ranges and 0 metadata columns:
#>           start       end     width
#>       <integer> <integer> <integer>
#>   [1]         1        16        16
#>   [2]         5        10         6
#>   [3]         1         0         0
#>   [4]         1         6         6
#>   [5]         1         8         8

cutRseq(subject, Rpattern)

#>   A DNAStringSet instance of length 5
#>     width seq
#> [1]    15 ATCATGCCATCATGA
#> [2]    10 CATGATATTA
#> [3]     0 
#> [4]     6 AAAAAA
#> [5]     3 AGG

cutLseq(subject, Lpattern, anchored = FALSE)

#>   A DNAStringSet instance of length 5
#>     width seq
#> [1]     2 AT
#> [2]     6 ATATTA
#> [3]     0 
#> [4]     6 AAAAAA
#> [5]     0 

cutLseq(subject, Lpattern, method = 'er', error_rate = 0.2)

#>   A DNAStringSet instance of length 5
#>     width seq
#> [1]    16 ATCATGCCATCATGAT
#> [2]    10 CATGATATTA
#> [3]     0 
#> [4]     6 AAAAAA
#> [5]     8 AGGTCATG

cutLseq(subject, Lpattern, method = 'er', error_rate = 0.2,
with.indels = TRUE)

#>   A DNAStringSet instance of length 5
#>     width seq
#> [1]    11 GCCATCATGAT
#> [2]    10 CATGATATTA
#> [3]     0 
#> [4]     6 AAAAAA
#> [5]     8 AGGTCATG